requirements.txt
pyserial
numpy
matplotlib
PyQt5